[feature #21]: Refactored pytest fixtures
