+++ tests/routers/v1/conftest.py
@@ -4,20 +4,11 @@
 
 """pytest fixtures for unit / integration tests."""
 
-from unittest.mock import AsyncMock, MagicMock
-
 import argon2
 import pytest
 from nmtfast.settings.v1.schemas import AuthApiKeySettings, SectionACL
-from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.v1.settings import AppSettings, AuthSettings, LoggingSettings
-from app.repositories.v1.gadgets import GadgetRepository
-from app.repositories.v1.widgets import WidgetRepository
-from app.schemas.v1.gadgets import GadgetRead, GadgetZapTask
-from app.schemas.v1.widgets import WidgetRead, WidgetZapTask
-from app.services.v1.gadgets import GadgetService
-from app.services.v1.widgets import WidgetService
 
 ph = argon2.PasswordHasher()
 
@@ -70,130 +61,3 @@ def mock_allow_acls() -> list[SectionACL]:
         SectionACL(section_regex=".*", permissions=["*"]),
     ]
     return acls
-
-
-@pytest.fixture
-def mock_async_session() -> AsyncMock:
-    """
-    Fixture to provide a mock AsyncSession.
-    """
-    return AsyncMock(spec=AsyncSession)
-
-
-@pytest.fixture
-def mock_gadget():
-    """
-    Fixture for a mock gadget dict (as stored in MongoDB).
-    """
-    return {
-        "id": "id-1",
-        "name": "ZapBot 5000",
-        "force": 5,
-    }
-
-
-@pytest.fixture
-def mock_mongo_db(mock_gadget):
-    """
-    Fixture for a mock MongoDB database with a 'gadgets' collection using AsyncMock.
-    """
-    collection = MagicMock()
-    collection.find_one = AsyncMock(return_value=mock_gadget.copy())
-    collection.update_one = AsyncMock(return_value=None)
-
-    mongo_db = {"gadgets": collection}
-    return mongo_db
-
-
-#
-# widget fixtures
-#
-
-
-@pytest.fixture
-def mock_widget_repository(mock_async_session: AsyncMock) -> WidgetRepository:
-    """
-    Fixture to provide a mock WidgetRepository.
-    """
-    return WidgetRepository(mock_async_session)
-
-
-@pytest.fixture
-def mock_widget_service(
-    mock_widget_repository: WidgetRepository,
-    mock_allow_acls: list[SectionACL],
-    mock_settings: AppSettings,
-) -> WidgetService:
-    """
-    Fixture to provide a mock WidgetService.
-    """
-    return WidgetService(mock_widget_repository, mock_allow_acls, mock_settings)
-
-
-@pytest.fixture
-def mock_widget_read() -> WidgetRead:
-    """
-    Fixture to provide a test WidgetRead instance.
-    """
-    return WidgetRead(id=1, name="Test Widget", height="10", mass="5", force=20)
-
-
-@pytest.fixture
-def mock_widget_zap_task() -> WidgetZapTask:
-    """
-    Fixture to provide a test WidgetZapTask instance.
-    """
-    return WidgetZapTask(
-        uuid="uuid-here",
-        state="PENDING",
-        id=1,
-        duration=1,
-        runtime=0,
-    )
-
-
-#
-# gadget fixtures
-#
-
-
-@pytest.fixture
-def mock_gadget_repository(mock_mongo_db: AsyncMock) -> GadgetRepository:
-    """
-    Fixture to provide a mock GadgetRepository.
-    """
-    return GadgetRepository(mock_mongo_db)
-
-
-@pytest.fixture
-def mock_gadget_service(
-    mock_gadget_repository: GadgetRepository,
-    mock_allow_acls: list[SectionACL],
-    mock_settings: AppSettings,
-) -> GadgetService:
-    """
-    Fixture to provide a mock GadgetService.
-    """
-    return GadgetService(mock_gadget_repository, mock_allow_acls, mock_settings)
-
-
-@pytest.fixture
-def mock_gadget_read() -> GadgetRead:
-    """
-    Fixture to provide a test GadgetRead instance.
-    """
-    return GadgetRead(id="id-1", name="Test Gadget", height="10", mass="5", force=20)
-
-
-@pytest.fixture
-def mock_gadget_zap_task() -> GadgetZapTask:
-    """
-    Fixture to provide a test GadgetZapTask instance.
-    """
-    return GadgetZapTask(
-        uuid="uuid-here",
-        state="PENDING",
-        id="id-1",
-        duration=1,
-        runtime=0,
-    )
